backend/demo/requirements.txt
flask==2.3.3
flask-cors==4.0.0
openai-whisper==20231117
cohere==4.37
yt-dlp==2023.11.16
werkzeug==2.3.7
torch>=2.0.0
transformers>=4.30.0
faiss-cpu>=1.7.4
numpy>=1.24.0
huggingface-hub>=0.16.0